+++ src/Resource/EnvelopeCategory.php
@@ -8,4 +8,5 @@ final class EnvelopeCategory extends BaseResource
 {
     public string $id;
     public string $name;
+    public string $value;
 }